Add list all books method
