+++ models/label.rb
